refactor(wpml): move reference remapping into Helpers::remapReference()
Extract the ACF reference-id → target-language remapping out of
WpmlBlockOverride into a shared Helpers::remapReference() primitive on the
field-formatting layer, alongside formatLink()'s existing WPML link
rewriting. WpmlBlockOverride now delegates Copy-field reference remapping
instead of carrying its own private remapReference/remapIds/remapId trio,
so the block-sync path and any field formatter resolve translated entities
the same way.

Behavior is unchanged: the per-field-type element-type resolution
(attachment / post type / taxonomy) and the non-reference / non-numeric
pass-through rules are identical. The remap test suite moves with the code
to tests/Unit/Helpers and now exercises the public method directly, no
reflection.
