+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Added
 
+- **`Parisek\TimberKit\Helpers::remapReference()`** — shared formatting-layer primitive that remaps an ACF reference field's id(s) to a target WPML language via `wpml_object_id`, with the element type resolved per field type (image / file / gallery → `attachment`; post_object / relationship / page_link → the referenced post's own type; taxonomy → the field's taxonomy). `WpmlBlockOverride` now delegates Copy-field reference remapping here instead of carrying its own private copy, so the block-sync path and any field formatter resolve translated entities the same way. Non-reference and non-numeric values pass through unchanged.
 - **`Parisek\TimberKit\WpmlBlockOverride` class** — runtime override of Copy field values in ACF Gutenberg blocks for WPML-multilingual sites. Hooks `render_block_data` at priority 20 and, for ACF blocks rendered in a non-default language, overwrites `attrs.data.<field>` for fields marked `wpml_cf_preferences = 1` (Copy) with the source-language post's value. Attachment IDs (image / file / gallery) are remapped to per-language duplicates via `wpml_object_id`. Supports nested Copy fields inside repeater / group containers at arbitrary depth via recursive `walkFields()` + path-aware key generation (e.g. `steps_N_image`, `faq_sections_N_items_M_title`). Cached as a single block-name → copy-fields index transient with per-request memo; persistent layer bypassed under `WP_DEBUG`. Filters exposed: `timber_kit/wpml_block_override/should_override`, `timber_kit/wpml_block_override/copy_fields`. Solves the long-standing WPML pain point where changing a Copy field (typically an image) in the source language never propagates to translated `post_content` without manual ATE re-job. Reference implementation in [portadesign/atelier99#14](https://github.com/portadesign/atelier99/pull/14); research, prior art, and design discussion in [#29](https://github.com/parisek/timber-kit/issues/29).
 
 ### Documentation
